pyproject.toml
[tool.poetry]
name = "be-ifinance"
version = "0.1.0"
description = "iFinance"
authors = ["arifrusmandani <arif.rusmandani@aqi.co.id>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.12"
uvicorn = "^0.34.2"
fastapi-utils = "^0.8.0"
typing-inspect = "^0.9.0"
sqlalchemy = "^2.0.40"
sqlalchemy-mixins = "^2.0.5"
psycopg2 = "^2.9.10"
httpx = "^0.28.1"
google-generativeai = "^0.8.5"
google-genai = "^1.11.0"
pydantic = {extras = ["email"], version = "^2.11.3"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.4.0"}
python-multipart = "^0.0.20"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
